Ensure paths and customPackages for component factory creator come from single source
