+++ packages/sitecore-jss-dev-tools/src/templating/nextjs/constants.ts
@@ -1,6 +1,3 @@
-// export outside of package
-export const COMPONENT_ROOT_PATH = 'src/components';
-export const PROJECT_ROOT_PATH = 'src/projects';
 // use only within the package
 export const projectsOutputPath = 'src/temp/projects';
 export const componentFactoryOutputPath = 'src/temp/componentFactoryCreator.ts';
