+++ packages/sitecore-jss-dev-tools/src/templating/nextjs/index.ts
@@ -1,6 +1,4 @@
-import * as constants from './constants';
 export { generateProjectComponents } from './templates/project-components';
 export { generateComponentFactoryCreator } from './templates/component-factory-creator';
-export { constants };
 export { writeComponentFactoryCreator } from './write-component-factory-creator';
 export { getProjectList } from './component-factory-utils';
